requirements.txt
django >= 2.0
django-tables2
django-filter
django-bootstrap3
pyopenssl >= 18
cryptography >= 2.2.1
